services/cv/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.206
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.4
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
sentry-sdk[fastapi]==1.38.0